feat: initial implementation of aigw CLI (#439)
**Commit Message**

This commit adds the initial implementation of aigw CLI.

As an initial functionality, this implements the static translator of
Envoy AI Gateway resources into Envoy Gateway and Kubernetes resources.
The implementation is partially done and follow up PR will improve the
functionality. The implementation uses the fake Kubernetes client and
reuses the controller code as-is. That allows us to avoid
re-implementing the translation logic twice, which could easily drift
and would be hard to maintain.



**Related Issues/PRs (if applicable)**

contributes to #412

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>