controller: update rotate function to return expiration time (#455)
**Commit Message**

Update `Rotate` to return the expiration time in addition to the error.
This will allow us to get the expiration time without having to make a
`GET` API request to Kubernetes.



**Related Issues/PRs (if applicable)**


https://github.com/envoyproxy/ai-gateway/blob/8f38c672ee11e6e724fff7b1e430979417fd3ce6/internal/controller/backend_security_policy.go#L106-L108

```
2025-03-02T20:12:23Z    INFO    controller.backend-security-policy      successfully rotated credentials for credentials in namespace default of auth type AWSCredentials, renewing in -0.009733 minutes
```

The requeue time after rotating a set of credentials (AWS in this case)
returns a negative number of minutes occasionally. It requires a manual
action/global sync to restart the rotation logic.

Upon further investigation, the issue found lies with
`rotator.GetPreRotationTime` returning the incorrect time.
GetPreRotation gets the expiration time from the secret's annotation,
and the assumption is it's getting a cached previous secret's annotation
instead of the new expiration time.


https://github.com/envoyproxy/ai-gateway/blob/8f38c672ee11e6e724fff7b1e430979417fd3ce6/internal/controller/backend_security_policy.go#L143-L147

---------

Signed-off-by: Aaron Choo <achoo30@bloomberg.net>
Co-authored-by: Dan Sun <dsun20@bloomberg.net>