deps: use latest EG, pin cel-go to 0.22 (#460)
**Commit Message**

The CLI's standalone mode needs to use Envoy Gateway's latest patch [1]
that allows us to use it inside our CLI. This upgrades the EG dependency
version to the latest. It is not compatible with the latest cel-go, so
this also pins it to 0.22 until the upstream k8s.io/apiserver dependency
[2] issue is resolved. This commit is separated from the standalone
implementation commit as the upgrade results in a CRD generation which
is not relevant to the change.

1: https://github.com/envoyproxy/gateway/pull/5407
2: https://github.com/kubernetes/apiserver/issues/116

**Related Issues/PRs (if applicable)**

This is a preparation for #458

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>