controller: refactor rotation interface (#477)
**Commit Message**
Refactor the rotation interface to make it easier to add other OIDC
rotation like Azure and GCP. The main change is to move the oidc token
token into the rotate function as we may have different token provider
such as Azure token service which needs to call azure specific APIs, we
do not want leak this code into the main reconciliation loop. Also we
removed the access token cache which is unnecessary as the call is only
made whenever the credential needs to be refreshed and we have timeout
to not block the controller.

---------

Signed-off-by: Dan Sun <dsun20@bloomberg.net>