extproc: allow extra attributes in the metrics interface (#484)
**Commit Message**

https://github.com/envoyproxy/ai-gateway/pull/484

**Related Issues/PRs (if applicable)**

Follow-up of: https://github.com/envoyproxy/ai-gateway/pull/483

**Special notes for reviewers (if applicable)**

This complements the previous pull request by making the
ChatCompletionMetrics capable of receiving additional attributes to be
added to the generated metrics. This will facilitate custom builds of
the completion metrics to enrich metrics with custom attributes.

Note that this would not be immediately leverageable by the default
build of extprox, as the chat completion Processor will call the methods
with no extra attirbtues, but it facilitates writing cleaner code when
building a custom implementation of the metrics. In addition, we will
have an interface that already makes sense and is better suited for
extension (and if we eventually allow users to customizer metrics via
the filter API config, etc, the metrics interface would already be able
to nicely accommodate that).

---------

Signed-off-by: Ignasi Barrera <ignasi@tetrate.io>