cli: make translator retain original resources (#481)
**Commit Message**

This changes the translator so that it will retain the non-AI Gateay
original resources and write them back into the output. This also fixes
the AI Gateway Route controller so that it can support the anonymous
resources that will be necessary in the translator as the EG standalone
mode requires the anonymous resources.

**Related Issues/PRs (if applicable)**

Extracted from #458

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>