extproc: fixes router to make the first match win (#510)
**Commit Message**

Previously, there was a bug in router's code where the matching rule
search doesn't break and it iterated over all rules even when there's a
matching rule found in the middle. That was a bug in the sense that it
is not only an inefficient but also semantically wrong since the last
match wins as opposed to the specification of GW API [1] which specifies
"If ties still exist within an HTTPRoute, matching precedence MUST be
granted to the FIRST matching rule (in list order) with a match meeting
the above criteria.".

1:
https://gateway-api.sigs.k8s.io/reference/spec/#gateway.networking.k8s.io%2fv1.HTTPRouteRule

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>