cli: initial implementation of `aigw run` (#458)
**Commit Message**

This implements the `aigw run` command that runs the AI Gateway locally.
This utilizes the Envoy Gateway's standalone mode [1]. With this
command, users are now able to use the AI Gateway, without Docker or
Kubernetes. The follow up PRs to further modify the default
configuration to add more de-facto services such as ollama, and will add
the command to show the default configuration.

1:
https://gateway.envoyproxy.io/docs/tasks/operations/standalone-deployment-mode/

**Related Issues/PRs (if applicable)**

Contributes to #412 
Follow up on #498

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>