feat: initial implementation of Inference Extension  (#493)
**Commit Message**

This commit scaffolds the foundation for the Inference Extension API
[1]. The design documentation was merged in #492. The controller needs
to be started with `--enableInferenceExtension=true` to not break the
existing controller deployment where the Inference Extension CRDs are
not installed.

This commit doesn't implement the actual "metrics-aware" load balancing
and instead it just does the random routing out of given (resolved)
endpoints. The follow up implementations will add more advanced
algorithm while expanding the metrics interface that currently only
provides the setter APIs.

The summary of the implementation is:
* Added `kind` field to AIGatewayRouteRuleBackendRef so that it can
reference InferencePool.
* InferencePool.Spec.Selector is allowed to specify multiple
AIServiceBackend.
* When building up all the extproc config via filterapi.Config, the
controller reads the referenced InferencePool and its binding
InferenceModels, and group them together into a single
filterapi.DynamicLoadBalancing configuration.
* When the extproc loads the configuration containing
DynamicLoadBalancing, it will resolve all the IP addresses for hostnames
belonging to the DyanmicLoadbalancing. The presence of
DynamicLoadBalancing in th config forces the config watcher reload and
refresh the config regardless of the updates. That way, the list of ip
addresses will always be updated (eventual consistency anyways) in a
non-hot path.
* On the request path, the ChatCompletionProcessor will check the
existence of the DynamicLoadBalancing config for the backend selected by
the router. If so, it further tries to resolve the ip:port level
endpoint selection.
* The selected ip:port will be set to the special header that will be
routed to ORIGINAL_DST.
* ORIGINAL_DST cluster will be added by the EG extension sever
implementation. Also, the extension server modifies some routes to
properly route to that cluster.

1: https://github.com/kubernetes-sigs/gateway-api-inference-extension

**Related Issues/PRs (if applicable)**

Built on #492
Contributes to #423

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>