controller: surface IsNotFound error in GetPreRotationTime (#552)
**Commit Message**

This commit ensures that an `apierrors.IsNotFound` is not discarded in
the `GetPreRotationTime` method.
Instead, the error is returned to the caller. This commit also updates
the `executeRotation` method to
maintain the preexisting behavior by allowing the `apierrors.IsNotFound`
case trigger the token rotation.

**Related Issues/PRs (if applicable)**

Fixes #466 

**Special notes for reviewers (if applicable)**

This commit does not add a new test for the `executeRotation` method as
the `TestBackendSecurityPolicyController_ReconcileOIDC_Fail` already
depends on this behavior.

Signed-off-by: John Landa <jonathanlanda@gmail.com>