refactor: stop sending request bodies twice to extproc (#636)
**Commit Message**

Previously, request body phase was unnecessarily called for upstream
filter extproc, hence the entire request body was traveling from Envoy
to the extproc at least twice (+ retry). This does the small refactoring
of the extproc code so that it will utilize CONTINUE_AND_REPLACE status
on the request headers phase since it can access the original body saved
in memory at the routing phase. This helps the memory pressure
reduction.

**Related Issues/PRs (if applicable)**

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>