feat: add ownedBy and createdAt fields to AIGatewayRoute (#620)
**Commit Message**

This commit added new fields CreatedAt & OwnedBy to AIGatewayRoute as
well as the internal Backend, so people can customize the output of
OpenAI-compatible APIs.

We set the fields at `AIGatewayRouteRule` level, the main two reasons
here:

- I believe for most of the cases, we only match one model at the
`Matches` fields, so working at `AIGatewayRouteRule` seems not bad as
the model:modelOwner = 1:1
- for different backends, the modelOwner:backends = 1:n, because
backends just means different route paths for one model, so working at
`AIGatewayRouteRule` level also makes sense.

One shortcoming here is in fallover mode, like we set different backends
for one model service, e.g. Envoy AI Gateway + openai, then we will have
the same modelOwner for them, it's a bit confusing.

But I persist on this just because when querying the `/models` list,
let's assume that we'll have two models sharing the same modelOwner
(different with the current implementation):

- even the enduser gets two models, but actually they can not select
which backend to query, the only knob for them is the model name
- the fallover backend, the openai here may never be queried, as a
result of this, return it to the user seems useless

**Related Issues/PRs (if applicable)**

Fixes: https://github.com/envoyproxy/ai-gateway/issues/588

**Special notes for reviewers (if applicable)**

None

---------

Signed-off-by: kerthcet <kerthcet@gmail.com>