test: deflake e2e by forwarding to svc, not pod (#660)
**Commit Message**

Previously, the pod was the target for port-forwarding. However, that
can be fragile since the pod may be rolled out even after the gateway
itself becomes healthy after the side car injection. This will deflake
the tests by using the service as a target which allows us to avoid
choosing the pod that is in the middle of termination/rollout.

**Related Issues/PRs (if applicable)**

Closes #647

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>