feat: add priority to backendRef (#673)
**Commit Message**

Add priority field to AIGatewayRouteRuleBackendRef. This will set
priority on the underlying LocalityLB endpoint in Envoy.

**Related Issues/PRs (if applicable)**

N/A

**Special notes for reviewers (if applicable)**

In order to get the Envoy Gateway to reconcile the underlying
`HTTPRoute` when the priorities in the `AIGatewayRoute` spec change we
need to add an annotation. This annotation just store the priorities for
each rule. This ensure we only trigger reconciliation when the
priorities actually change.

---------

Signed-off-by: Curtis Maddalozzo <cmaddalozzo@bloomberg.net>
Signed-off-by: Dan Sun <dsun20@bloomberg.net>
Co-authored-by: Dan Sun <dsun20@bloomberg.net>