extproc: hotfix for gzip header (#683)
**Commit Message**

This is a hotfix for a bug where if the gzip header is present, we
decompress and leave the header. This results in the client throwing an
error trying to decompress based on the header, but it has already been
decompressed. This hot fix removes the header. In the future we probably
want to return it compressed with the header intact.

for example - the error I encountered was
```io.netty.handler.codec.compression.DecompressionException: Input is
not in the GZIP format```

Co-authored-by: Sukumar Gaonkar <sgaonkar4@bloomberg.net>

---------

Signed-off-by: Alexa Griffith <agriffith96@gmail.com>
Co-authored-by: Sukumar Gaonkar <sgaonkar4@bloomberg.net>