feat: add model name field to backend ref (#687)
**Commit Message**

Add model name field to backend ref. This allows users pass a different
model name to the backend from the one that was provided in the original
request. This is particularly useful for cross-provided failover and
load balancing.

For example: If a user has provisioned Claude 3.5 Sonnet on both AWS and
GCP they can setup a rule that will load balance across them:

```yaml
apiVersion: aigateway.envoyproxy.io/v1alpha1
kind: AIGatewayRoute
metadata:
  name: test-route
  namespace: ai-gateway
spec:
  targetRefs:
  - group: gateway.networking.k8s.io
    kind: Gateway
    name: eg
  schema:
    name: OpenAI
  rules:
  - matches:
      - headers:
        - type: Exact
          name: x-ai-eg-model
          value: claude-3-5-sonnet-v2
    backendRefs:
    - name: aws-backend
    - name: gcp-backend
```

Since the both providers will have a different name/model identifier for
the provisioned endpoint the above configuration will not work. This PR
makes it possible to add a `modelName` on the backend ref which
specifies the providers name for the model.

With the `modelName` property we can support this use-case as follows:

```yaml
apiVersion: aigateway.envoyproxy.io/v1alpha1
kind: AIGatewayRoute
metadata:
  name: test-route
  namespace: ai-gateway
spec:
  targetRefs:
  - group: gateway.networking.k8s.io
    kind: Gateway
    name: eg
  schema:
    name: OpenAI
  rules:
  - matches:
      - headers:
        - type: Exact
          name: x-ai-eg-model
          value: claude-3-5-sonnet-v2
    backendRefs:
    - name: aws-backend
      modelName: arn:aws:bedrock:us-west-2:123456789012:provisioned-model/abc123xyz
    - name: gcp-backend
      modelName: claude-3-5-sonnet-v2@20241022
```

---------

Signed-off-by: Curtis Maddalozzo <cmaddalozzo@bloomberg.net>
Signed-off-by: Curtis Maddalozzo <cmaddalozzo@users.noreply.github.com>
Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>
Co-authored-by: Takeshi Yoneda <t.y.mathetake@gmail.com>