controller: defaults to /v1 for OpenAI when empty version (#765)
**Commit Message**

Previously, the controller didn't fallback to /v1 version for OpenAI
schema since #706 which starts utilizing schema.version field. The field
was previously non nil hence in many cases, it was set to empty, and
that caused the backward incompatibility. This fixes it by defaulting to
v1 when the field is set but empty string.

**Related Issues/PRs (if applicable)**

Follow up on #706

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>
Signed-off-by: ydu208 <YDU208@bloomberg.net>