fix: add content-length on request body transformation (#818)
**Description**

We use REPLACE_AND_CONTINUE in the request header phase on the upstream
extproc filter. However, the current extproc implementation removes the
content-length header unconditionally when body mutation happens with
REPLACE_AND_CONTINUE flag.

This adds a workaround of that limitation by inserting the header
mutation filter after the upstream extproc filter where we retrieve the
content-length from a dynamic metadata.

**Related Issues/PRs (if applicable)**

Fixes the issue that the #730 resolves in different way

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>