extproc: return 404 instead of 500 for unknown path (#835)
**Description**

Previously, unknown path was responded as an internal error as opposed
to the fact that it's an 404 with the user input root cause. This fixes
the extproc code that way, now that users will be able to know what's
wrong with the operation instead of getting the cryptic 500 error.

**Related Issues/PRs (if applicable)**

Contributes to #810
Closes #724

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>