test: adds real provider embeddings test & update doc (#841)
**Description**

This adds embeddings endpoint tests with the providers that support the
endpoint. This only added the providers for which we have credentials.
According to the testing situation we have right now, this also
clarifies in the "Supported Endpoints" page that which provider is
tested and which is not for each endpoint.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>