extproc: add content-length metadata regardless of mutation (#915)
**Description**

Previously, content-length dynamic metadata was only set when the
request body mutation happens. That is because we assumed that the
content-length is always present when not. However, it seems that some
epp removes the content-length header because of the way it works, so
this will add a content-length header regardless of the mutation.

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>