ci: separate out precommit job into a different workflow (#918)
**Description**

Previously, the precommit check was only running on the code changes
since it exists inside the build_and_test.yaml workflow since #774. This
was a problematic when, for example, the PR only changing the
documentation was opened because the spellchecks etc were not running on
that PR. This essentially reverts #774.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>