cli: fixes race with EG logs for aigw run (#945)
**Description**

Previously, there was a flake in the use of bytes.Buffer passed to EG
command as well as on our aigw run side, we are reading it concurrently.
This fixes it temporarily by passing os.Stdout or io.Discard.

The race was occasionally happening on CI which was causing the flake,
so this fixes it.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>