extproc: ignore nanoseconds part of created timestamp (#962)
**Description**

Previously, "created" field in the chat completion responses etc assumes
that its value is a json number without floating point part, meaning it
must be parsable as int64 at Go level. Since OpenAI clearly states that
it is an integer json number in the API documentation, it is natural
logic we have. However, Sambanova returns the created field in the
1234.56789 format where it includes "nanoseconds" in the floating point
part which results in the parsing error at the extproc. This fixes it by
ignoring the nanoseconds part since we are not using it on the response
path anyways.


**Related Issues/PRs (if applicable)**

Closes #959

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>