controller: eliminate EEP in favor of extension server logic (#963)
**Description**

This switches the logic to insert the router/listener-level AIGW extproc
filter. Previously, we did by generating EnvoyExtensionPolicy (EEP).
However, only one EEP can be attached per Gateway, so it was difficult
for users to insert their own policy in conjunction with AIGW (e.g.
#590, #903).

This commit eliminates the EEP generation and switched to the logic
where we insert the filter through the extension server implementation.
This allows us to fine-control exactly where we enable the AIGW extproc,
hence naturally resolves the issue such as supporting the normal
non-AIGW-generated HTTPRoutes to be attached to the same gateway without
affecting these routes #940.

Note that this is an internal breaking change. Upgrading AIGW to the
version including this commit will not be affected by change. However,
users are encouraged to migrate to the new logic by manually deleting
the EEP previously generated by AIGW.

**Related Issues/PRs (if applicable)**

Closes #590
Closes #903
Closes #836
Closes #940

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>
Co-authored-by: Xunzhuo <bitliu@tencent.com>