test: openinference span recording from python (#972)
**Description**

This creates the tests which gate completion of tracing support (#791).
These are intentionally disabled as functionality isn't yet written, but
raising this as a separate part allows smaller code review in the
future, and also allows others to participate without blocking (e.g. by
skipping only certain cassettes, e.g. at first all but "chat-basic")

**Related issues**

#791

**Special notes for reviewers (if applicable)**

It is important that we know what the span data expectations are, hence
the testopeninference package. This was one of the most difficult things
to figure out, so I hope this helps. docker won't run unless we
re-record the spans. Also, in the future, we can pin the python
requirements.txt for a specific openinference version to know what we
are aiming to be compatible with.

cc @mikeldking maybe something like this could eventually go upstream,
granted that we need a copy here for parity with the cassettes we are
recording, because the cassettes are used for reasons besides otel. food
for thought.

Signed-off-by: Adrian Cole <adrian@tetrate.io>