feat: add configurable HTTP request headers to metrics (#981)
**Description**

This commit adds support for configuring additional metrics labels
corresponding to request headers. Notably, this adds a new flag to the
controller `metricsRequestHeaderLabelMapping` that will be validated and
propagated to the extproc.

The flag is in the format of the comma separated key value pairs as in
"header1:label1,header2:label2,..."

Fixes #848

---------

Signed-off-by: Suren Raju <suren.raju@careem.com>
Co-authored-by: Takeshi Yoneda <t.y.mathetake@gmail.com>