extproc: remove input api schema in filterapi config (#998)
**Description**

This removes the input api schema field in the filter config as it
shouldn't be a config and currently it is not really used. The input
schema is already determined by the specific API endpoint being used by
the Envoy AI Gateway, such as the unified OpenAI API at
`v1/chat/completions` or later the Anthropic Messages API we are
supporting at `anthropic/v1/messages`. This removes the confusion when
we are going to support other vendor pass through APIs like Anthropic.
We may need to consider deprecating the input `APISchema` defined on
`AIGatewayRoute` which is also not really used.

Related Issue: #847

---------

Signed-off-by: Dan Sun <dsun20@bloomberg.net>
Co-authored-by: Takeshi Yoneda <t.y.mathetake@gmail.com>