controller: adds support for GatewayNamespace mode (#988)
**Description**

Previously, we assumed that Envoy Gateway was deployed in the standard
mode where Gateway pods, deployments and daemonsets are all deployed and
managed in the Envoy Gateway namespace. However, there's an emerging new
feature of Envoy Gateway, called "GatewayNamespace" [1], where these
resources are created locally in the same namespace as Gateway object.

This change generalizes Gateway controller's logic so that AIGW works
regardless of the Envoy Gateway deployment model. As a result, we no
longer need `--envoyGatewayNamespace` flag.


[1]:
https://gateway.envoyproxy.io/docs/tasks/operations/gateway-namespace-mode/

**Related Issues/PRs (if applicable)**

Closes #976

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>