extproc: chat completion tracing support for all backends (#1068)
**Description**

This commit is a follow up on the initial tracing implementation. There
was a bug in vcr tests where streaming mode override is wrongly taken
place at upstream filter level, which resulted in buffering all events
at the router filter as opposed to the reality that they are streamed
(not necessarily on the event boundary, as it's a tcp dataframe level),
and that is why previously already passing the tests despite not taking
care of the event boundaries.

Notably, this makes the tracing do
* Support for all backends regardless of the schema vs previously only
works for OpenAI backend.
* Correctly handling all streaming events vs previously assuming all
events are available in buffer which is not true in end to end.
* Avoids unnecessary copies around the tracing; This utilizes the
already-parsed data structures, so this makes the tracing zero-copy as
long as the redaction is configured.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>