test: fix bad recording of openai embeddings spans (#1140)
**Description**

This fixes the test data that was unnoticed as bad even though the tests
passed.

What happens is canonically all openai SDKs backfill an embedding
`encoding_format`, so our requests didn't add that and this caused a
request mismatch. `encoding_format` is backfilled to `base_64` when
platform support exists for it, in python for example. This fixes the
test fixture which silently succeeded and re-recorded all the data.

**Related Issues/PRs (if applicable)**

#696

---------

Signed-off-by: Adrian Cole <adrian@tetrate.io>