fix: adds upgrade testing & graceful termination workaround (#1248)
**Description**

This adds an end-to-end upgrade testing where we are supposed to test
two test scenarios: simply rolling upgrade and control plane upgrade
while keep making requests and verify no requests are dropped. What we
found is that, as reported in #1241, there's a slight gap between Envoy
stop receiving requests and extproc termination, hence users might
experience requests being dropped during upgrade.

The fundamental fix is to set extproc sidecar container in the k8s API
sense, but it's only available after k8s v1.33 by default. So, this adds
a common workaround to sleep before the context cancelation. The
workaround fix is verified to work in the newly added e2e upgrade tests.

After this is merged, adding k8s-version detection and enabling sidecar
by default automatically as well as backporting the fix to v0.3 would be
necessary. After that, we can enable the control-plane upgrade variant
of the test case that is currently disabled in this commit.

**Related Issues/PRs (if applicable)**

Closes #1241
Closes #1060

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>